@@ -13,29 +13,30 @@
     "lint:format": "prettier  --loglevel warn --write \"./**/*.{js,jsx,css,md,json}\" --ignore-path ./.prettierignore"
   },
   "dependencies": {
-    "@vitejs/plugin-react": "^3.0.0",
-    "compression": "^1.7.4",
-    "cross-env": "^7.0.3",
-    "express": "^4.18.1",
+    "@vitejs/plugin-react": "4.2.0",
+    "compression": "1.7.4",
+    "cross-env": "7.0.3",
+    "express": "4.18.2",
     "react": "18.2.0",
     "react-content-loader": "6.2.1",
     "react-dom": "18.2.0",
-    "react-intl": "6.4.2",
-    "react-toggle-dark-mode": "^1.1.1",
-    "sirv": "^2.0.2",
-    "vite": "^4.0.3",
-    "vite-plugin-ssr": "^0.4.131",
-    "zustand": "^4.3.8"
+    "react-intl": "6.5.5",
+    "react-toggle-dark-mode": "1.1.1",
+    "sirv": "2.0.3",
+    "vike": "0.4.146",
+    "vite": "5.0.0",
+    "zustand": "4.4.6"
   },
   "devDependencies": {
-    "autoprefixer": "10.4.14",
-    "eslint": "8.41.0",
-    "eslint-config-prettier": "8.8.0",
-    "eslint-plugin-jsx-a11y": "6.7.1",
-    "eslint-plugin-prettier": "4.2.1",
-    "eslint-plugin-react": "7.32.2",
+    "autoprefixer": "10.4.16",
+    "eslint": "8.53.0",
+    "eslint-config-prettier": "9.0.0",
+    "eslint-plugin-jsx-a11y": "6.8.0",
+    "eslint-plugin-prettier": "5.0.1",
+    "eslint-plugin-react": "7.33.2",
     "pre-commit": "1.2.2",
-    "prettier": "2.8.8"
+    "prettier": "3.1.0",
+    "vite-plugin-svgr": "^4.2.0"
   },
   "browserslist": [
     "> 0.2%",
@@ -43,5 +44,6 @@
     "not dead"
   ],
   "pre-commit": "lint",
-  "license": "MIT"
+  "license": "MIT",
+  "type": "module"
 }
