@@ -1,8 +1,13 @@
-const express = require('express')
-const compression = require('compression')
-const { renderPage } = require('vite-plugin-ssr/server')
+import express from 'express'
+import compression from 'compression'
+import sirv from 'sirv'
+import path from 'path'
+import { fileURLToPath } from 'url'
+
+import { renderPage } from 'vike/server'
 
 const isProduction = process.env.NODE_ENV === 'production'
+const __dirname = path.dirname(fileURLToPath(import.meta.url))
 const root = `${__dirname}/..`
 
 startServer()
@@ -13,7 +18,6 @@ async function startServer() {
   app.use(compression())
 
   if (isProduction) {
-    const sirv = require('sirv')
     app.use(sirv(`${root}/dist/client`))
   } else {
     const vite = require('vite')
@@ -33,10 +37,12 @@ async function startServer() {
     const pageContext = await renderPage(pageContextInit)
     const { httpResponse } = pageContext
     if (!httpResponse) return next()
-    const { body, statusCode, contentType, earlyHints } = httpResponse
+    const { body, statusCode, headers, earlyHints } = httpResponse
     if (res.writeEarlyHints)
       res.writeEarlyHints({ link: earlyHints.map((e) => e.earlyHintLink) })
-    res.status(statusCode).type(contentType).send(body)
+    res.status(statusCode)
+    headers.forEach(([name, value]) => res.setHeader(name, value))
+    res.send(body)
   })
 
   const port = process.env.PORT || 3000
