@@ -1,13 +1,13 @@
-import { FormattedMessage } from 'react-intl'
 import styles from '#/styles/BlockQuote.module.css'
 import Section from '#/components/Section'
+import Text from '#/components/Text'
 
 function BlockQuote({ quote, author }) {
   return (
     <Section as='blockquote' className={styles.base}>
       <span>
         <span className={styles.quote}>“</span>
-        <FormattedMessage id={quote} />
+        <Text id={quote} />
         <span className={styles.quote}>”</span>
       </span>
       <div className={styles.author}>— {author}</div>
