@@ -1,26 +1,34 @@
 import styles from '#/styles/Navbar.module.css'
-
+import { useEffect } from 'react'
 import { DarkModeSwitch } from 'react-toggle-dark-mode'
 
-import useTheme from '#/hooks/useTheme'
 import useAppStore from '#/store'
 
-import logo from '#/assets/svg/logo.svg'
+import Icon from '#/components/Icon'
+import LogoIcon from '#/assets/svg/logo.svg?react'
 import { handleScrollToTop } from '#/utilities/scroll'
 
+const verifyTheme = (theme) => {
+  if (theme === 'dark') document.querySelector('body').classList.add('dark')
+  else document.querySelector('body').classList.remove('dark')
+}
+
 function Navbar() {
-  const { isDarkTheme, changeTheme } = useTheme()
-  const { locale, toggleLocale } = useAppStore()
+  const { locale, theme, toggleLocale, toggleTheme } = useAppStore()
+
+  useEffect(() => {
+    verifyTheme(theme)
+  }, [theme])
 
   return (
     <nav className={styles.base}>
       <div className={styles.items}>
-        <button className={styles.logo} onClick={handleScrollToTop}>
-          <img src={logo} alt='Logotipo' />
+        <button type='button' className={styles.logo} onClick={handleScrollToTop}>
+          <Icon icon={LogoIcon} />
         </button>
         <DarkModeSwitch
-          checked={isDarkTheme}
-          onChange={changeTheme}
+          checked={theme === 'dark'}
+          onChange={toggleTheme}
           moonColor='var(--content-color)'
           sunColor='var(--content-color)'
         />
