@@ -1,20 +1,16 @@
 import styles from '#/styles/ProjectsSection.module.css'
-import { FormattedMessage } from 'react-intl'
 
 import Section from '#/components/Section'
 import Project from '#/components/Project'
+import Text from '#/components/Text'
 
 import projects from '#/constants/projects'
 
 function ProjectsSection() {
   return (
     <Section>
-      <h2>
-        <FormattedMessage id='projects.title' />
-      </h2>
-      <p>
-        <FormattedMessage id='projects.text1' />
-      </p>
+      <Text as={'h2'} id='projects.title' />
+      <Text as={'p'} id='projects.text1' />
       <div className={styles.projects}>
         {projects.map((project, i) => (
           <Project key={i} {...project} />
