@@ -1,17 +1,15 @@
 import styles from '#/styles/SkillsSection.module.css'
-import { FormattedMessage } from 'react-intl'
 
 import Section from '#/components/Section'
 import Skill from '#/components/Skill'
 
 import skills from '#/constants/skills'
+import Text from '#/components/Text'
 
 function SkillsSection() {
   return (
     <Section>
-      <h2>
-        <FormattedMessage id='skills.title' />
-      </h2>
+      <Text as={'h2'} id='skills.title' />
       {Object.entries(skills).map(([title, specificSkills]) => (
         <div className={styles.subsection} key={title}>
           <h3>{title}</h3>
